Create interface for GameController, and create explicit game state
