+++ app/build.gradle
@@ -28,4 +28,5 @@ dependencies {
     implementation files('libs/snakeyaml-1.19-android.jar')
     compileOnly 'org.projectlombok:lombok:1.16.20'
     annotationProcessor 'org.projectlombok:lombok:1.16.20'
+    compile 'com.android.support:appcompat-v7:26.0.0-beta1'
 }
